+++ almost_make/utils/makeUtil.py
@@ -42,7 +42,7 @@ class MakeUtil:
 
     def __init__(self):
         self.macroCommands["shell"] = lambda code, macros: os.popen(self.macroUtil.expandMacroUsages(code, macros)).read().rstrip(' \n\r\t') # To-do: Use the built-in shell if specified...
-        self.macroCommands["wildcard"] = lambda argstring, macros: " ".join([ shlex.quote(part) for part in globber.glob(self.macroUtil.expandMacroUsages(argstring, macros), '.') ])
+        self.macroCommands["wildcard"] = lambda argstring, macros: " ".join([ shlex.quote(part) for part in self.glob(self.macroUtil.expandMacroUsages(argstring, macros), macros) ])
         self.macroCommands["words"] = lambda argstring, macros: str(len(SPACE_CHARS.split(self.macroUtil.expandMacroUsages(argstring, macros))))
         self.macroCommands["sort"] = lambda argstring, macros: " ".join(sorted(list(set(SPACE_CHARS.split(self.macroUtil.expandMacroUsages(argstring, macros))))))
         self.macroCommands["strip"] = lambda argstring, macros: argstring.strip()
@@ -137,6 +137,81 @@ class MakeUtil:
         targetNames.extend(specialTargetNames)
         return (result, targetNames)
 
+    # Get a list of directories (including the current working directory)
+    # from macros['VPATH']. Returns an array with one element, the current working
+    # directory, if there is no 'VPATH' macro.
+    def getSearchPath(self, macros):
+        searchPath = [ os.path.abspath('.') ]
+        
+        if not 'VPATH' in macros:
+            return searchPath
+
+        vpath = macros['VPATH']
+
+        # Split first by ';', then by ':', then finally,
+        # try to split by space characters.
+        splitOrder = [';', ':', ' ']
+        split = []
+        for char in splitOrder:
+            split = escaper.escapeSafeSplit(vpath, char, True)
+            split = runner.removeEmpty(split)
+
+            if len(split) > 1:
+                break
+        
+        searchPath.extend([ os.path.normcase(part) for part in split ])
+
+        return searchPath
+
+    # Find a file with relative path [givenPath]. If 
+    # VPATH is in macros, search each semi-colon, colon,
+    # or space-separated entry for the file. Returns the 
+    # path to the file, or None, if the file does not exist.
+    def findFile(self, givenPath, macros):
+        givenPath = os.path.normcase(givenPath)
+        
+        searchPath = self.getSearchPath(macros)
+
+        for part in searchPath:
+            path = os.path.join(part, givenPath)
+
+            if os.path.exists(path):
+                return path
+        return None
+
+    # Glob [text], but search [VPATH] for additional matches.
+    def glob(self, text, macros):
+        if not 'VPATH' in macros:
+            return globber.glob(text, '.')
+        
+        searchPath = self.getSearchPath(macros)
+        result = globber.glob(text, '.', [])
+        text = os.path.normcase(text)
+
+        for part in searchPath:
+            result.extend(globber.glob(os.path.join(part, text), '.', []))
+        
+        # Act like system glob. If we didn't find anything, 
+        # return [ text ] 
+        if len(result) == 0:
+            result = [ text ]
+        
+        return result
+
+    # Glob all elements in arr, but not the first.
+    def globArgs(self, arr, macros, excludeFirst=True):
+        result = []
+        isFirst = excludeFirst
+
+        for part in arr:
+            if not runner.isQuoted(part.strip()) and not isFirst:
+                result.extend(self.glob(part, macros))
+            else:
+                result.append(part)
+                isFirst = False
+        
+        return result
+
     # Generate [target] if necessary. Returns
     # True if generated, False if not necessary.
     def satisfyDependencies(self, target, targets, macros):
@@ -171,7 +246,7 @@ class MakeUtil:
                     if len(parts) > 2:
                         continue
                     
-                    if not targets[".SUFFIXES"]:
+                    if not ".SUFFIXES" in targets:
                         continue
                     
                     validSuffixes,_ = targets[".SUFFIXES"]
@@ -192,7 +267,9 @@ class MakeUtil:
                         
                         targets[target] = (newDeps, rules)
                         break
-        selfExists = os.path.exists(target)
+        
+        targetPath = self.findFile(target, macros)
+        selfExists = targetPath != None
         selfMTime = 0
 
         if not target in targets:
@@ -201,11 +278,12 @@ class MakeUtil:
             else:
                 self.errorUtil.reportError("No rule to make %s." % target)
                 return True # If still running, the user wants us to exit successfully.
-        runRecipe = False
+        
         deps, commands = targets[target]
+        runRecipe = False
         
         if selfExists:
-            selfMTime = os.path.getmtime(target)
+            selfMTime = os.path.getmtime(targetPath)
 
         def isPhony(target):
             if not ".PHONY" in targets:
@@ -213,28 +291,43 @@ class MakeUtil:
             
             phonyTargets,_ = targets['.PHONY']
             return target in phonyTargets or target in MAGIC_TARGETS
+        
         selfPhony = isPhony(target)
         
         def needGenerate(other):
-            return isPhony(other) \
-                or not os.path.exists(other) \
-                or selfMTime > os.path.getmtime(other) \
-                or not selfExists \
-                or selfPhony
+            if isPhony(other) or selfPhony or not selfExists:
+                return True
 
-        if isPhony(target):
+            pathToOther = self.findFile(other, macros)
+
+            return pathToOther == None # or selfMTime > os.path.getmtime(pathToOther)
+
+        if selfPhony:
             runRecipe = True
 
+        if not selfExists:
+            runRecipe = True
+        
+        depPaths = []
+        deps = self.globArgs(runner.removeEmpty(deps), macros, False) # Glob the set of dependencies.
+
+        for dep in deps:
+            if isPhony(dep):
+                depPaths.append(dep)
+            pathToDep = self.findFile(dep, macros)
+
+            depPaths.append(pathToDep or dep)
+
         if not runRecipe:
-            if not selfExists:
-                runRecipe = True
-            else:
-                for dep in deps:
-                    if isPhony(dep) or \
-                        not os.path.exists(dep) \
-                        or os.path.getmtime(dep) >= selfMTime:
-                            runRecipe = True
-                            break
+            for dep in depPaths:
+                if isPhony(dep):
+                    runRecipe = True
+                    break
+
+                if os.path.getmtime(dep) > selfMTime:
+                    runRecipe = True
+                    break
+        
         # Generate each dependency, if necessary.
         if not runRecipe:
             return False
@@ -272,10 +365,10 @@ class MakeUtil:
         # (1) all dependencies are satisfied
         # (2) we need to run each command in recipe.
         # Define several macros the client will expect here:
-        macros["@"] = target
-        macros["^"] = " ".join(deps)
+        macros["@"] = targetPath or target
+        macros["^"] = " ".join(depPaths)
         if len(deps) >= 1:
-            macros["<"] = deps[0]
+            macros["<"] = depPaths[0]
 
         for command in commands:
             command = self.macroUtil.expandMacroUsages(command, macros).strip()
@@ -335,7 +428,7 @@ class MakeUtil:
                 parts = runner.shSplit(line)
                 command = parts[0].strip()
 
-                parts = runner.globArgs(parts, runner.ShellState()) # Glob all, except the first...
+                parts = self.globArgs(parts, macros) # Glob all, except the first...
                 parts = parts[1:] # Remove leading include...
                 ignoreError = False
 
@@ -345,6 +438,12 @@ class MakeUtil:
 
                 for fileName in parts:
                     fileName = runner.stripQuotes(fileName)
+
+                    if not os.path.exists(fileName):
+                        foundName = self.findFile(fileName, macros)
+
+                        if foundName != None:
+                            fileName = foundName
 
                     if not os.path.exists(fileName):
                         if ignoreError: